Split out a new misc integration test tag
