+++ tests/integration/appdash_test.go
@@ -1,6 +1,9 @@
 // Copyright 2016-2021, Pulumi Corporation.  All rights reserved.
 //
 // Utilities for testing AppDash-based tracing files.
+//
+//go:build misc || all
+// +build misc all
 
 package ints
